WIP of my testbed-focused potion proxy dockerfile
There's some stuff in here specifically for debugging, but I figured
since no one else is using it it's nice to at least push it up for
posterity's sake for the moment. Once the new release is fixed, the bit
where I use a new .tar.gz for the release should go away.